CVSL-2391: Move Visited enum.
