+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsassessforearlyreleaseapi/service/policy/model/residentialchecks/ResidentialChecksPolicy.kt
@@ -66,3 +66,8 @@ data class ResidentialChecksPolicy(
   val version: PolicyVersion,
   val tasks: List<Task>,
 )
+
+enum class VisitedAddress {
+  I_HAVE_VISITED_THIS_ADDRESS_AND_SPOKEN_TO_THE_MAIN_OCCUPIER,
+  I_HAVE_NOT_VISITED_THE_ADDRESS_BUT_I_HAVE_SPOKEN_TO_THE_MAIN_OCCUPIER,
+}
